test: colocate tests and rule documentation
Move all four Vitest suites into import.meta.vitest blocks beside the implementation they exercise. Configure each package to discover in-source tests, provide the ImportMeta types, and replace import.meta.vitest during package builds so test-only dependencies are removed from published output.

Move both rule Markdown documents verbatim into file-level JSDoc and point rule metadata at the shipped TypeScript sources. This keeps examples, options, rationale, and references next to each implementation without separate documentation files.
