+++ packages/oxlint/tsconfig.json
@@ -12,8 +12,8 @@
 		"skipLibCheck": true,
 		"strict": true,
 		"target": "ES2024",
-		"types": ["node", "vitest/globals"],
+		"types": ["node", "vitest/importMeta"],
 		"verbatimModuleSyntax": true
 	},
-	"include": ["src", "test", "vite.config.ts"]
+	"include": ["src", "vite.config.ts"]
 }
